added timestamp to sensor config

+++ src/ruuvi_driver_sensor.h
@@ -110,7 +110,7 @@ typedef struct __attribute__ ( (packed, aligned (4)))
     uint8_t dsp_parameter;  //!< Parameter to DSP functions
     uint8_t mode;           //!< Mode, RD_SENSOR_SLEEP, _SINGLE, _CONTINOUS
     uint8_t divider;      //!< Reserved for future use
-    uint16_t created_at;      //!< Reserved for future use
+    uint64_t created_at;      //!< Reserved for future use
 }
 rd_sensor_configuration_t;
